requirements.txt
altair==5.1.2
ansi2html==1.8.0
anyio==3.6.2
apispec==6.3.0
appdirs==1.4.4
appnope==0.1.3
argcomplete==1.10.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
astroid==2.14.1
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
blis==0.7.9
boto3==1.26.89
botocore==1.29.89
branca==0.6.0
brotlipy==0.7.0
build==0.10.0
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.5.18.1
chardet==3.0.4
click==8.1.3
comm==0.1.2
compressed-rtf==1.0.6
confection==0.0.4
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
coverage==7.1.0
cycler==0.11.0
cymem==2.0.7
Cython==3.0.2
dash==2.13.0
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
Django==4.2.10
django-rest-swagger==2.2.0
djangorestframework==3.14.0
dnspython==2.3.0
docopt==0.6.2
docutils==0.19
docx2txt==0.8
drf-yasg==1.21.7
ebcdic==1.1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
extract-msg==0.28.7
fastjsonschema==2.16.2
filelock==3.12.0
flake8==6.0.0
Flask==2.2.3
Flask-Dance==6.2.0
Flask-Login==0.6.2
flask-swagger-ui==4.11.1
Flask-WTF==1.1.1
folium==0.14.0
fonttools==4.38.0
fqdn==1.5.1
frozendict==2.3.8
gitdb==4.0.11
GitPython==3.1.40
google==3.0.0
google-auth==2.16.2
google-search-results==2.4.2
googlefinance==0.7
html5lib==1.1
IMAPClient==2.1.0
importlib-metadata==6.0.0
inflection==0.5.1
iniconfig==2.0.0
invoke==1.7.1
ipykernel==6.20.2
ipympl==0.9.3
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
itypes==1.2.0
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.5
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.12.0
keyring==23.13.1
kiwisolver==1.4.4
langcodes==3.3.0
lazy-object-proxy==1.9.0
llvmlite==0.40.1
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.4
more-itertools==9.1.0
moto==4.1.4
mpmath==1.3.0
multitasking==0.0.11
murmurhash==1.0.9
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.2
notebook-as-pdf==0.5.0
notebook_shim==0.2.2
numba==0.57.1
numpy==1.24.1
oauthlib==3.2.2
olefile==0.46
openapi-codec==1.3.2
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.1
patsy==0.5.3
pdfminer.six==20191110
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.11
pkginfo==1.9.6
platformdirs==2.6.2
plotly==5.17.0
pluggy==1.0.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.25.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycosat==0.6.3
pycryptodome==3.17
pydantic==1.10.6
pydeck==0.8.1b0
pyee==8.2.2
pyflakes==3.0.1
Pygments==2.14.0
pylint==2.16.1
pymongo==4.3.3
pyparsing==3.0.9
pypdf==3.5.2
PyPDF2==3.0.1
pyppeteer==1.0.2
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytesseract==0.3.10
pytest==7.2.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.4
python-pptx==0.6.21
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.0
readme-renderer==37.3
regex==2022.10.31
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.23.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.3.5
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.8.0
serpapi==0.1.5
simplejson==3.19.2
six==1.12.0
sklearn==0.0
smart-open==6.3.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.8.1
sqlparse==0.4.4
srsly==2.4.6
stack-data==0.6.2
statsmodels==0.13.5
streamlit==1.28.1
sympy==1.11.1
tenacity==8.2.3
terminado==0.17.1
tesseract==0.1.3
textblob==0.17.1
textract==1.6.5
thinc==8.1.9
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.2
traitlets==5.8.1
twine==4.0.2
typer==0.7.0
types-PyYAML==6.0.12.8
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
uri-template==1.2.0
uritemplate==4.1.1
URLObject==2.4.3
validators==0.22.0
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.14.1
WTForms==3.0.1
xlrd==1.2.0
XlsxWriter==3.1.0
xmltodict==0.13.0
yarg==0.1.9
yfinance==0.2.20
zipp==3.11.0
